ranking-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.13.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
aioredis==2.0.1
loguru==0.7.2
prometheus-client==0.19.0
PyYAML==6.0.1
pytest==7.4.3
pytest-asyncio==0.21.1